Moved JSON interfaces to dataHandler folder
